requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
django-jazzmin==2.4.9
django-parler==2.2.1
django-parler-rest==2.1
django-redis==5.0.0
django-summernote==0.8.20.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
celery==5.1.2
ClickUz==0.3
googletrans==4.0.0rc1
gunicorn==20.1.0
PaycomUz==3.0
Pillow==9.1.1
redis==3.5.3
requests==2.26.0
drf-yasg==1.20.0
sorl-thumbnail==12.7.0
django-rosetta==0.9.8
psycopg2-binary==2.9.3
django-resized==1.0.0
django-smartfields==1.1.3